python-vuln-demo/requirements.txt
#
# Intentionally vulnerable dependency pins for testing security scanners.
# Do not use these versions in production.
#
requests==2.19.1
PyYAML==5.3.1
Jinja2==2.11.2
